♻️ refactor: Base presigned URL 코드 코어로 분리
